% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquiring_modules.R
\name{loadExam}
\alias{loadExam}
\title{loadExam}
\usage{
loadExam(exam_file, type = c("bioc", "bm", "gen", "biol"))
}
\arguments{
\item{exam_file, }{a character string indicating the path and the excel file containing the exam results}

\item{type, }{a character string indicating the exam type. Four options: bioc 9 credits, bm 5 credits, gen 3 credits and biol 1 credit}
}
\value{
It returns a data.frame with the information of the students which have passed a specific module
}
\description{
This function is designed to create a data.frame where, given a specific module, a normalised score is generated.
}
\examples{
\dontrun{
bioc = loadExam(exam_file=paste(getwd(), "bioc.xls",sep="/"), type="bioc")
bm = loadExam(exam_file=paste(getwd(), "bm.xls",sep="/"), type="bm")
g = loadExam(exam_file=paste(getwd(), "g.xls",sep="/"), type="gen")
biol = loadExam(exam_file=paste(getwd(), "biol.xls",sep="/"), type="biol")
}
}
\author{
Raffaele Calogero
}
