% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_students_db.R
\name{mkDB}
\alias{mkDB}
\title{mkDB}
\usage{
mkDB(new_students, studentDB = NULL, updateDB = FALSE)
}
\arguments{
\item{new_students, }{a character string indicating the path of excel file containing the new students}

\item{studentDB, }{a character string indicating the path of csv file containing the previous set of students}

\item{updateDB, }{a logical FALSE, this is a new db to be built from scratch and save in studentDB, TRUE update the db at studentDB, this is done every new year whne new students are incoming}
}
\description{
This function is is design to create a students csv file with the 4 score initialised as NA.
}
\examples{
\dontrun{
mkDB(
  new_students = "/the/input/dir/file.csv",
  studentDB = "studentsdb.csv",
  updateDB = T
)
# first DB
mkDB(
  new_students = paste(getwd(), "1anno_Canali_ABCD_25-26.csv", sep="/"),
  updateDB = F
)
}
}
\author{
Raffaele Calogero
}
