% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populating_db.R
\name{populatingDB}
\alias{populatingDB}
\title{populatingDB}
\usage{
populatingDB(norm_score, studentDB, type = c("bioc", "bm", "gen", "biol"))
}
\arguments{
\item{norm_score, }{the name of the variable contaning the normalised score generated by loadExam.}

\item{studentDB, }{a data.frame of the LATEST update of studentDB: i.e. including last verbalising exam and, every year, the new students entering in the second semester of the first year}

\item{type, }{a character string indicating the exam type. Four options: bioc 9 credits, bm 5 credits, gen 3 credits and biol 1 credit}
}
\value{
It returns a student database populated with the results of the students modules
}
\description{
This function is designed to populate the studentDB with the normalised score from a specific module. IMPORTANT: the studentDB file must the the one referring to the previous verbalising exam
}
\examples{
\dontrun{
db = utils::read.csv("somewhere/BandB/studentDB_2026-07-18.csv") #reading the latest studentDB
db = populatingDB(norm_score = bioc, studentDB = db, type="bioc")
db = populatingDB(norm_score = bioc, studentDB = db, type="bm")
db = populatingDB(norm_score = bioc, studentDB = db, type="gen")
db = populatingDB(norm_score = bioc, studentDB = db, type="biol")
utils::write.csv(db, paste0("/somewhere/BandB//studentDB_", format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M-\%S"),".csv"), row.names = F) #writing the updated studentDB
}
}
\author{
Raffaele Calogero
}
