% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbalising.R
\name{scoring}
\alias{scoring}
\title{scoring}
\usage{
scoring(registered_students, studentDB)
}
\arguments{
\item{registered_students, }{a character string indicating the path and the excel file containing the students registered for the final scoring}

\item{studentDB, }{a character string indicating the path of the ladtes csv file containing students and results}
}
\value{
It returns a comma separated file verbalising_tmp.csv with the names of students to be verbalised and a file called verbalising_to_be_manually_revised.csv in case there are students with less than 17 credits to be handled manually.
}
\description{
This function is designed to create a students csv file with the 4 score initialised as NA.
}
\examples{
\dontrun{
scoring(
  registered_students = "/the/input/dir/verbalizzante.xls",
  studentDB = "/somewhere/studentDB_2026-07-15_18-49-07.csv"
)
}
}
\author{
Raffaele Calogero
}
